+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 #Change Log
 All notable changes to this project starting with the 0.6.0 release will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org)
 
-## MASTER
+## [0.10.6] - 2016-03-07
 ### Changed
 - `Terminus\Helpers\AuthHelper` has become `Terminus\Models\Auth`. (#971)